+++ CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+- Added properties for accessing the current average fps and milliseconds per frame.
+- Changed stats_fontsize command to also affect the fps display.
 - Fixed stats created via the custom attribute not restoring display status between sessions.
 - Fixed font size issue in console input for commands with a long signature.