+++ CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+
+## [1.0.4] - 2021-12-09
 - Added properties for accessing the current average fps and milliseconds per frame.
 - Changed cs_evaluate to also work without the command name being specified.
 - Changed stats_fontsize command to also affect the fps display.